relocate EditCourseButton
